# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Basket Order Module",
    "version": "1.0.0",
    "description": "Go port of the supplied Java module. See docs/PARITY.md for source defects, sample conflicts, and integration requirements. Business failures generally use HTTP 200 and status Not ok."
  },
  "servers": [
    {
      "url": "http://localhost:9009"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/basketorder/add/scrips": {
      "post": {
        "operationId": "addBasketScrip",
        "tags": [
          "basket"
        ],
        "description": "Enabled by modules.basket. Adds one scrip; maximum business.max_scrips defaults to 20. Returns all basket scrips. Exchange/token/symbol/expiry/format/week-tag are refreshed from the contract cache; add response lotSize is null.\n\nRequest fields:\n- basketId (integer; Required): Positive ID owned by the caller.\n- scrips (Scrip; Required): See shared Scrip fields below. A single object, not an array.\n\nBusiness messages: Invalid Parameter, Invalid basket, Basket reached the maximum limits",
        "responses": {
          "200": {
            "description": "Response (check status for business failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "example": {
                  "status": "Ok",
                  "message": "Success",
                  "result": [
                    {
                      "id": 2,
                      "basketId": 2,
                      "lotSize": null,
                      "sortOrder": null,
                      "exchange": "NSE",
                      "token": "2188",
                      "tradingSymbol": "GENCON-EQ",
                      "qty": "1",
                      "price": "43.04",
                      "expiry": null,
                      "product": "CNC",
                      "transType": "BUY",
                      "priceType": "MKT",
                      "orderType": "Regular",
                      "ret": "DAY",
                      "triggerPrice": "",
                      "disClosedQty": null,
                      "mktProtection": null,
                      "target": "0",
                      "stopLoss": "0",
                      "trailingStopLoss": "",
                      "formattedInsName": "GENCON-EQ",
                      "weekTag": null,
                      "validityDays": null,
                      "expiryDate": null
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Route disabled or absent"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "basketId": {
                    "type": "integer",
                    "description": "Positive ID owned by the caller. Requirement: Required."
                  },
                  "scrips": {
                    "$ref": "#/components/schemas/ScripRequestModel",
                    "description": "See shared Scrip fields below. A single object, not an array. Requirement: Required."
                  }
                },
                "required": [
                  "basketId",
                  "scrips"
                ]
              },
              "example": {
                "basketId": 2,
                "scrips": {
                  "exchange": "NSE",
                  "token": "2188",
                  "tradingSymbol": "GENCON-EQ",
                  "qty": "1",
                  "price": "43.04",
                  "product": "CNC",
                  "transType": "BUY",
                  "priceType": "MKT",
                  "orderType": "Regular",
                  "ret": "DAY",
                  "source": "WEB"
                }
              }
            }
          }
        },
        "summary": "Add one scrip",
        "externalDocs": {
          "description": "Complete API reference",
          "url": "API_REFERENCE.md#op-addscrip"
        }
      }
    },
    "/basketorder/create": {
      "post": {
        "operationId": "createBasketName",
        "tags": [
          "basket"
        ],
        "description": "Enabled by modules.basket. Creates the basket and returns the user\u2019s personal basket list, newest ID first.\n\nRequest fields:\n- basketName (string; Required): Nonblank; unique for the authenticated user.\n\nBusiness messages: Invalid Parameter, Basket name already exist",
        "responses": {
          "200": {
            "description": "Response (check status for business failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "example": {
                  "status": "Ok",
                  "message": "Success",
                  "result": [
                    {
                      "basketId": 2,
                      "basketName": "Smoke personal",
                      "createdOn": 1789033567235,
                      "isExecuted": "0",
                      "scripCount": 0
                    },
                    {
                      "basketId": 1,
                      "basketName": "Expired fixture",
                      "createdOn": 1789033567233,
                      "isExecuted": "0",
                      "scripCount": 1
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Route disabled or absent"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "basketName": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "Nonblank; unique for the authenticated user. Requirement: Required."
                  }
                },
                "required": [
                  "basketName"
                ]
              },
              "example": {
                "basketName": "Smoke personal"
              }
            }
          }
        },
        "summary": "Create a personal basket",
        "externalDocs": {
          "description": "Complete API reference",
          "url": "API_REFERENCE.md#op-create"
        }
      }
    },
    "/basketorder/delete/scrips": {
      "post": {
        "operationId": "deleteBasketScrip",
        "tags": [
          "basket"
        ],
        "description": "Enabled by modules.basket. Deletes matching scrips and returns the remaining scrip array, which can be empty.\n\nRequest fields:\n- basketId (integer; Required): Positive owned basket ID.\n- scripsId (integer[]; Required): Nonempty list of scrip IDs belonging to this basket.\n\nBusiness messages: Invalid Parameter, Invalid basket, No records found",
        "responses": {
          "200": {
            "description": "Response (check status for business failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "example": {
                  "status": "Ok",
                  "message": "Success",
                  "result": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Route disabled or absent"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "basketId": {
                    "type": "integer",
                    "description": "Positive owned basket ID. Requirement: Required."
                  },
                  "scripsId": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "description": "Nonempty list of scrip IDs belonging to this basket. Requirement: Required."
                  }
                },
                "required": [
                  "basketId",
                  "scripsId"
                ]
              },
              "example": {
                "basketId": 2,
                "scripsId": [
                  2
                ]
              }
            }
          }
        },
        "summary": "Delete selected basket scrips",
        "externalDocs": {
          "description": "Complete API reference",
          "url": "API_REFERENCE.md#op-deletescrip"
        }
      }
    },
    "/basketorder/delete/{basketId}": {
      "delete": {
        "operationId": "deleteBasket",
        "tags": [
          "basket"
        ],
        "description": "Enabled by modules.basket. Requires ownership. Deletes the basket and its scrips transactionally; related notifications are deleted when the notification module is enabled. Returns the remaining personal basket list.\n\nBusiness messages: Invalid Parameter, Invalid basket",
        "responses": {
          "200": {
            "description": "Response (check status for business failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "example": {
                  "status": "Ok",
                  "message": "Success",
                  "result": [
                    {
                      "basketId": 1,
                      "basketName": "Expired fixture",
                      "createdOn": 1789033567233,
                      "isExecuted": "0",
                      "scripCount": 1
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Route disabled or absent"
          }
        },
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Delete a personal basket",
        "externalDocs": {
          "description": "Complete API reference",
          "url": "API_REFERENCE.md#op-delete"
        }
      }
    },
    "/basketorder/execute": {
      "post": {
        "operationId": "excuteBasketOrder",
        "tags": [
          "basket"
        ],
        "description": "Enabled by modules.basket. Submits the request\u2019s scrips to the configured bulk-order service, forwarding the bearer token. Does not load the saved basket scrips as the order request. A non-null upstream response marks the basket executed; individual broker rejections are not aggregated. Returns an ARRAY of response envelopes. No automatic order retry.\n\nRequest fields:\n- basketId (integer; Required): Positive owned basket ID.\n- scrips (Scrip[]; Required): Nonempty submitted order list; tradingSymbol and source required per scrip.\n\nBusiness messages: Invalid Parameter, Invalid basket, Failed",
        "responses": {
          "200": {
            "description": "Response (check status for business failures)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenericResponse"
                  }
                },
                "example": [
                  {
                    "status": "Ok",
                    "message": "Basket executed successfully",
                    "result": null
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Route disabled or absent"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "basketId": {
                    "type": "integer",
                    "description": "Positive owned basket ID. Requirement: Required."
                  },
                  "scrips": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ScripRequestModel"
                    },
                    "description": "Nonempty submitted order list; tradingSymbol and source required per scrip. Requirement: Required."
                  }
                },
                "required": [
                  "basketId",
                  "scrips"
                ]
              },
              "example": {
                "basketId": 2,
                "scrips": [
                  {
                    "exchange": "NSE",
                    "token": "2188",
                    "tradingSymbol": "GENCON-EQ",
                    "qty": "1",
                    "price": "43.04",
                    "product": "CNC",
                    "transType": "BUY",
                    "priceType": "MKT",
                    "orderType": "Regular",
                    "ret": "DAY",
                    "source": "WEB"
                  }
                ]
              }
            }
          }
        },
        "summary": "Execute a personal basket",
        "externalDocs": {
          "description": "Complete API reference",
          "url": "API_REFERENCE.md#op-execute"
        }
      }
    },
    "/basketorder/get": {
      "get": {
        "operationId": "getBasketNames",
        "tags": [
          "basket"
        ],
        "description": "Enabled by modules.basket. Returns basketId, basketName, isExecuted (string \"0\"/\"1\"), createdOn (epoch milliseconds), and scripCount. Research-call baskets are excluded. An empty list returns status Ok, message No records found, result null.",
        "responses": {
          "200": {
            "description": "Response (check status for business failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "example": {
                  "status": "Ok",
                  "message": "Success",
                  "result": [
                    {
                      "basketId": 2,
                      "basketName": "Smoke personal",
                      "createdOn": 1789033567235,
                      "isExecuted": "0",
                      "scripCount": 0
                    },
                    {
                      "basketId": 1,
                      "basketName": "Expired fixture",
                      "createdOn": 1789033567233,
                      "isExecuted": "0",
                      "scripCount": 1
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Route disabled or absent"
          }
        },
        "summary": "List personal baskets",
        "externalDocs": {
          "description": "Complete API reference",
          "url": "API_REFERENCE.md#op-get"
        }
      }
    },
    "/basketorder/get/scrips/{basketId}": {
      "get": {
        "operationId": "retrieveScrips",
        "tags": [
          "basket"
        ],
        "description": "Enabled by modules.basket. Requires basket ownership. Returns scrips ordered by ID and refreshes lotSize from the contract cache. No scrips: status Ok, message No records found, result null.\n\nBusiness messages: Invalid Parameter, Invalid basket",
        "responses": {
          "200": {
            "description": "Response (check status for business failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "example": {
                  "status": "Ok",
                  "message": "Success",
                  "result": [
                    {
                      "id": 2,
                      "basketId": 2,
                      "lotSize": "1",
                      "sortOrder": null,
                      "exchange": "NSE",
                      "token": "2188",
                      "tradingSymbol": "GENCON-EQ",
                      "qty": "1",
                      "price": "43.04",
                      "expiry": null,
                      "product": "CNC",
                      "transType": "BUY",
                      "priceType": "MKT",
                      "orderType": "Regular",
                      "ret": "DAY",
                      "triggerPrice": "",
                      "disClosedQty": null,
                      "mktProtection": null,
                      "target": "0",
                      "stopLoss": "0",
                      "trailingStopLoss": "",
                      "formattedInsName": "GENCON-EQ",
                      "weekTag": null,
                      "validityDays": null,
                      "expiryDate": null
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Route disabled or absent"
          }
        },
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Get basket scrips",
        "externalDocs": {
          "description": "Complete API reference",
          "url": "API_REFERENCE.md#op-scrips"
        }
      }
    },
...
    "/basketorder/reset/{basketId}": {
      "get": {
        "operationId": "resetExecutionStatus",
        "tags": [
          "basket"
        ],
        "description": "Enabled by modules.basket. Sets isExecuted to \"0\" for an owned basket and returns the personal basket list. This does not cancel broker orders.\n\nBusiness messages: Invalid Parameter, Invalid basket",
        "responses": {
          "200": {
            "description": "Response (check status for business failures)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                },
                "example": {
                  "status": "Ok",
                  "message": "Success",
                  "result": [
                    {
                      "basketId": 2,
                      "basketName": "Smoke renamed",
                      "createdOn": 1789033567235,
                      "isExecuted": "0",
                      "scripCount": 1
                    },
                    {
                      "basketId": 1,
                      "basketName": "Expired fixture",
                      "createdOn": 1789033567233,
                      "isExecuted": "0",
                      "scripCount": 1
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Route disabled or absent"
          }
        },
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "summary": "Reset basket execution flag",
        "externalDocs": {
          "description": "Complete API reference",
          "url": "API_REFERENCE.md#op-reset"
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
...
      "GenericResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "result"
        ],
        "properties": {
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {}
        }
      },
...
    }
  }
}